+++ baybe/parameters/custom.py
@@ -3,8 +3,10 @@
 from functools import cached_property
 from typing import Any, ClassVar, Union
 
+import numpy as np
 import pandas as pd
-from attr import define, field
+from attrs import define, field
+from attrs.validators import min_len
 
 from baybe.parameters.base import DiscreteParameter
 from baybe.parameters.enum import CustomEncoding
@@ -25,7 +27,7 @@ class CustomDiscreteParameter(DiscreteParameter):
     # See base class.
 
     # object variables
-    data: pd.DataFrame = field(eq=eq_dataframe)
+    data: pd.DataFrame = field(validator=min_len(2), eq=eq_dataframe)
     """A mapping that provides the encoding for all available parameter values."""
 
     decorrelate: Union[bool, float] = field(
@@ -48,27 +50,37 @@ class CustomDiscreteParameter(DiscreteParameter):
         """Validate the dataframe with the custom representation.
 
         Raises:
-            ValueError: If the dataframe contains ``NaN``.
-            ValueError: If the dataframe contains duplicated indices.
             ValueError: If the dataframe contains non-numeric values.
-            ValueError: If the dataframe contains columns that only contain a single
-                value.
+            ValueError: If the dataframe index contains non-string values.
+            ValueError: If the dataframe index contains empty strings.
+            ValueError: If the dataframe contains ``NaN``.
+            ValueError: If the dataframe index contains duplicates.
+            ValueError: If the dataframe contains columns with only one unique value.
         """
-        if value.isna().any().any():
-            raise ValueError(
-                f"The custom dataframe for parameter {self.name} contains NaN "
-                f"entries, which is not supported."
-            )
-        if len(value) != len(set(value.index)):
-            raise ValueError(
-                f"The custom dataframe for parameter {self.name} contains "
-                f"duplicated indices. Please only provide dataframes with unique"
-                f" indices."
-            )
         if value.select_dtypes("number").shape[1] != value.shape[1]:
             raise ValueError(
                 f"The custom dataframe for parameter {self.name} contains "
                 f"non-numeric values."
+            )
+        if not all(isinstance(x, str) for x in value.index):
+            raise ValueError(
+                f"The custom dataframe for parameter {self.name} contains non-string "
+                f"index values."
+            )
+        if not all(len(x) > 0 for x in value.index):
+            raise ValueError(
+                f"The custom dataframe for parameter {self.name} contains empty string "
+                f"index values."
+            )
+        if not np.isfinite(value.values).all():
+            raise ValueError(
+                f"The custom dataframe for parameter {self.name} contains nan/infinity "
+                f"entries."
+            )
+        if len(value) != len(set(value.index)):
+            raise ValueError(
+                f"The custom dataframe for parameter {self.name} contains "
+                f"duplicated indices."
             )
         if any(value.nunique() == 1):
             raise ValueError(