+++ tests/serialization/test_parameter_serialization.py
@@ -1,24 +1,37 @@
 """Test serialization of parameters."""
 
+import hypothesis.strategies as st
 import pytest
+from hypothesis import given
+from pytest import param
 
 from baybe.parameters.base import Parameter
 
+from ..hypothesis_strategies import (
+    categorical_parameter,
+    custom_parameter,
+    numerical_continuous_parameter,
+    numerical_discrete_parameter,
+    substance_parameter,
+    task_parameter,
+)
+
 
 @pytest.mark.parametrize(
-    "parameter_names",
+    "strategy",
     [
-        ["Categorical_1"],
-        ["Num_disc_1"],
-        ["Custom_1"],
-        ["Solvent_1"],
-        ["Conti_finite1"],
-        ["Task"],
+        param(numerical_discrete_parameter(), id="NumericalDiscreteParameter"),
+        param(numerical_continuous_parameter(), id="NumericalContinuousParameter"),
+        param(categorical_parameter(), id="CategoricalParameter"),
+        param(task_parameter(), id="TaskParameter"),
+        param(substance_parameter(), id="SubstanceParameter"),
+        param(custom_parameter(), id="CustomParameter"),
     ],
 )
-@pytest.mark.parametrize("n_grid_points", [5])
-def test_parameter_serialization(parameters):
-    param = parameters[0]
-    string = param.to_json()
-    param2 = Parameter.from_json(string)
-    assert param == param2
+@given(data=st.data())
+def test_parameter_roundtrip(strategy, data):
+    """A serialization roundtrip yields an equivalent object."""
+    parameter = data.draw(strategy)
+    string = parameter.to_json()
+    parameter2 = Parameter.from_json(string)
+    assert parameter == parameter2, (parameter, parameter2)